requirements.txt
lxml
doit
qgis-plugin-ci
wheel 
numpy 
cython 
setuptools
# PyGithub
# pylint>=1.9.4
mkdocs>=1.0
markupsafe==2.0.1
mkdocs-material>=3.3
python-markdown-math>=0.6
Jinja2==2.11.3
bleach>=3.1
pybtex>=0.24
mkdocs-bibtex>=2.16
# git+https://github.com/tramebleue/mdx_bib#egg=mdx_bib
